+++ src/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/GroupNormalizer.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\MongoDB\Normalizer;
 
-use Pim\Bundle\CatalogBundle\Entity\Group;
+use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Pim\Bundle\TransformBundle\Normalizer\Structured\TranslationNormalizer;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
@@ -43,6 +43,6 @@ class GroupNormalizer implements NormalizerInterface
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof Group && 'mongodb_json' === $format;
+        return $data instanceof GroupInterface && 'mongodb_json' === $format;
     }
 }