+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Allow to generate many versions in a single request
 - Better display on batch warnings
 - Add flat / csv denormalizers for product data
+- Introduce a GroupInterface and use it (to replace Entity\Group)
 
 ## BC breaks
 - Rename `Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\ResolveDoctrineOrmTargetEntitiesPass` to `Pim\Bundle\CatalogBundle\DependencyInjection\Compiler\ResolveDoctrineTargetModelsPass`
@@ -115,6 +116,7 @@
 - IndexCreator constructor now expects a LoggerInterface as last argument
 - Add methods isLocaleSpecific and getLocaleSpecificCodes in AttributeInterface
 - AssociationTransformer constructor now expects a $associationTypeClass as last argument
+- Inject the GroupFactory as las constructor argument in GroupController and VariantGroupController
 
 ## Bug fixes
 - PIM-3332: Fix incompatibility with overriden category due to usage of ParamConverter in ProductController